MonitoredItemCreateRequest
