+++ src/protocol/deserialize_auto.cpp
@@ -2,9 +2,9 @@
 // It is automatically generated from opcfoundation.org schemas.
 //
 
-/// @author Olivier Roulet-Dubonnet 
-/// @email olivier@sintef.no 
-/// @brief Opc Ua Binary. 
+/// @author Olivier Roulet-Dubonnet
+/// @email olivier@sintef.no
+/// @brief Opc Ua Binary.
 /// @license GNU LGPL
 ///
 /// Distributed under the GNU LGPL License
@@ -18,7 +18,7 @@
 #include <opc/ua/protocol/binary/stream.h>
 
 namespace OpcUa
-{   
+{
     namespace Binary
     {
 
@@ -2780,8 +2780,6 @@ namespace OpcUa
 
 */
 
-/*  DISABLED
-
     template<>
     void DataDeserializer::Deserialize<MonitoringParameters>(MonitoringParameters& data)
     {
@@ -2792,9 +2790,6 @@ namespace OpcUa
         *this >> data.DiscardOldest;
     }
 
-*/
-
-/*  DISABLED
 
     template<>
     void DataDeserializer::Deserialize<MonitoredItemCreateRequest>(MonitoredItemCreateRequest& data)
@@ -2804,7 +2799,6 @@ namespace OpcUa
         *this >> data.RequestedParameters;
     }
 
-*/
 
 /*  DISABLED
 
@@ -2820,19 +2814,14 @@ namespace OpcUa
 
 */
 
-/*  DISABLED
-
     template<>
-    void DataDeserializer::Deserialize<CreateMonitoredItemsParameters>(CreateMonitoredItemsParameters& data)
+    void DataDeserializer::Deserialize<MonitoredItemsParameters>(MonitoredItemsParameters& data)
     {
         *this >> data.SubscriptionId;
         *this >> data.TimestampsToReturn;
         DeserializeContainer(*this, data.ItemsToCreate);
     }
 
-*/
-
-/*  DISABLED
 
     template<>
     void DataDeserializer::Deserialize<CreateMonitoredItemsRequest>(CreateMonitoredItemsRequest& data)
@@ -2842,7 +2831,6 @@ namespace OpcUa
         *this >> data.Parameters;
     }
 
-*/
 
 /*  DISABLED
